+++ R/POSIXct2Character.R
@@ -47,7 +47,7 @@
 
 POSIXct2Character <- function(x, fmt="%Y-%m-%d %H:%M:%OS3") {
 
-  checkmate::assertCharacter(fmt, any.missing=FALSE, max.len=1)
+  checkmate::assertString(fmt)
 
   pos <- gregexpr("%OS[[:digit:]]+", fmt)[[1]]
   if (pos > 0) {